pligrim_bot/requirements.txt
aiofiles==25.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anitya_schema==2.3.0
annotated-types==0.7.0
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
constants==0.6.0
crochet==2.1.1
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
fedora-messaging-the-new-hotness-schema==1.2.1
fedora_messaging==3.9.0
frozenlist==1.8.0
google-auth==2.43.0
google-auth-oauthlib==1.2.3
gspread==6.2.1
gssapi==1.10.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
koji==1.35.3
magic-filter==1.0.12
multidict==6.7.0
oauthlib==3.3.1
pika==1.3.2
pillow==12.0.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
redis==7.0.1
referencing==0.37.0
requests==2.32.5
requests-gssapi==1.4.0
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
the-new-hotness==1.3.0
tomli==2.3.0
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==2.0.1
yarl==1.22.0
zope.interface==8.1.1
